refactor(engine-types): move transaction receipt into own module
