+++ dan_layer/engine_types/src/lib.rs
@@ -20,8 +20,8 @@ pub mod resource;
 pub mod resource_container;
 pub mod serde_with;
 pub mod substate;
+pub mod transaction_receipt;
 pub mod vault;
 
 mod template;
-
 pub use template::{calculate_template_binary_hash, TemplateAddress};
